+++ context.go
@@ -9,6 +9,7 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
+	"time"
 
 	"github.com/golang/dep/gps"
 	"github.com/golang/dep/internal/fs"
@@ -34,13 +35,14 @@ import (
 //	}
 //
 type Ctx struct {
-	WorkingDir     string      // Where to execute.
-	GOPATH         string      // Selected Go path, containing WorkingDir.
-	GOPATHs        []string    // Other Go paths.
-	Out, Err       *log.Logger // Required loggers.
-	Verbose        bool        // Enables more verbose logging.
-	DisableLocking bool        // When set, no lock file will be created to protect against simultaneous dep processes.
-	Cachedir       string      // Cache directory loaded from environment.
+	WorkingDir     string        // Where to execute.
+	GOPATH         string        // Selected Go path, containing WorkingDir.
+	GOPATHs        []string      // Other Go paths.
+	Out, Err       *log.Logger   // Required loggers.
+	Verbose        bool          // Enables more verbose logging.
+	DisableLocking bool          // When set, no lock file will be created to protect against simultaneous dep processes.
+	Cachedir       string        // Cache directory loaded from environment.
+	CacheAge       time.Duration // Maximum valid age of cached source data. <=0: Don't cache.
 }
 
 // SetPaths sets the WorkingDir and GOPATHs fields. If GOPATHs is empty, then
@@ -99,6 +101,7 @@ func (c *Ctx) SourceManager() (*gps.SourceMgr, error) {
 	}
 
 	return gps.NewSourceManager(gps.SourceManagerConfig{
+		CacheAge:       c.CacheAge,
 		Cachedir:       cachedir,
 		Logger:         c.Out,
 		DisableLocking: c.DisableLocking,