+++ gps/source_cache_bolt.go
@@ -21,6 +21,10 @@ import (
 	"github.com/pkg/errors"
 )
 
+// boltCacheFilename is a versioned filename for the bolt cache. The version
+// must be incremented whenever incompatible changes are made.
+const boltCacheFilename = "bolt-v1.db"
+
 // boltCache manages a bolt.DB cache and provides singleSourceCaches.
 type boltCache struct {
 	db     *bolt.DB
@@ -30,7 +34,7 @@ type boltCache struct {
 
 // newBoltCache returns a new boltCache backed by a BoltDB file under the cache directory.
 func newBoltCache(cd string, epoch int64, logger *log.Logger) (*boltCache, error) {
-	path := sourceCachePath(cd, "bolt") + ".db"
+	path := filepath.Join(cd, boltCacheFilename)
 	dir := filepath.Dir(path)
 	if fi, err := os.Stat(dir); os.IsNotExist(err) {
 		if err := os.MkdirAll(dir, os.ModeDir|os.ModePerm); err != nil {